# improve the perf for pruned array lookup cpu