Summary:
100x improvement :)

Lesson: We used the wrong tensor accessor here.

Before this Diff:
```
INFO:2022-08-09 16:45:04 2628811:2628811 CuptiActivityProfiler.cpp:158] CUDA versions. CUPTI: 14; Runtime: 11040; Driver: 11040
INFO:root:LinearTable: B: 2048, T: 100, L: 20, E: 100000, QPS: 0.00B QPS/s, T: 2400660us, Pruning Ratio: 90.00%, Table size: 0 GB
```

After this Diff:
```
BUILD SUCCEEDED
INFO:2022-08-09 18:31:06 3307940:3307940 CuptiActivityProfiler.cpp:158] CUDA versions. CUPTI: 14; Runtime: 11040; Driver: 11040
INFO:root:LinearTable: B: 2048, T: 100, L: 20, E: 100000, QPS: 0.17B QPS/s, T: 24758us, Pruning Ratio: 90.00%, Table size: 0 GB
```

Reviewed By: jiyuanzFB

Differential Revision: D38561337